requirements.txt
streamlit
openai-whisper
reportlab
pydub
langchain
langchain_chroma
langchain-text-splitters
langchain_community
python-dotenv
openai
datetime